+++ src/main/java/com/mx/edu/tecdesoftware/StreamCore/api/domain/service/PlanService.java
@@ -2,6 +2,8 @@ package com.mx.edu.tecdesoftware.StreamCore.api.domain.service;
 
 import com.mx.edu.tecdesoftware.StreamCore.api.domain.Plan;
 import com.mx.edu.tecdesoftware.StreamCore.api.domain.repository.PlanRepository;
+import com.mx.edu.tecdesoftware.StreamCore.api.domain.repository.SubscriptionRepository;
+import com.mx.edu.tecdesoftware.StreamCore.api.web.exception.ConflictException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -13,6 +15,9 @@ public class PlanService {
 
     @Autowired
     private PlanRepository planRepository;
+
+    @Autowired
+    private SubscriptionRepository subscriptionRepository;
 
     public List<Plan> getAll() {
         return planRepository.getAll();
@@ -28,6 +33,9 @@ public class PlanService {
 
     public boolean delete(int planId) {
         return getPlan(planId).map(plan -> {
+            if (subscriptionRepository.existsByPlan(planId)) {
+                throw new ConflictException("No se puede eliminar: el plan '" + planId + "' tiene suscripciones asociadas.");
+            }
             planRepository.delete(planId);
             return true;
         }).orElse(false);
